Front End Changes

+++ public/javascript/account.js
@@ -1,21 +1,20 @@
 class Account extends HTMLElement {
 
     static get observedAttributes() {
-        return ['events', 'points'];
-    }
-
-    clickaccount() {
-        console.log('this is a test');
+        return ['events', 'points', 'mode'];
     }
 
     events = ""
     points = ""
     name = ""
 
+    mode = 'DEVMODE'; // 'INTEGRATED'
+
     constructor() {
         // Always call super first in constructor
         super();
 
+        console.log('INITIALIZED ACCOUNT VIEW');
         var customElement = this;
 
         let template = document.getElementById('accountview');
@@ -37,7 +36,9 @@ class Account extends HTMLElement {
 
         let deleteButton = sr.getElementById("deleteAccountButton")
         deleteButton.addEventListener("click", e => {
-            this.delete();
+            if(MODE=='INTEGRATED'){
+                this.delete();
+            }
         })
     }
 
@@ -83,8 +84,9 @@ class Account extends HTMLElement {
 
         /* where to make a data call for points/events */
 
-        this.events = customElement.getAttribute('events')
-        this.points = customElement.getAttribute('points')
+        this.mode = customElement.getAttribute('mode');
+        this.events = customElement.getAttribute('events');
+        this.points = customElement.getAttribute('points');
 
         if (this.events == null) {
             this.events = '-'
@@ -98,40 +100,45 @@ class Account extends HTMLElement {
         if(this.name == null){
             this.name = "";
         }else{
-            console.log('name parameter passed in')
+            console.log('SETTING NAME')
         }
 
-        this.eventsattended = sr.getElementById('eventsattended');
-        this.eventsattended.innerHTML = this.events;
-        this.pointearned = sr.getElementById('pointearned');
-        this.pointearned.innerHTML = this.points;
+        
         this.nameelement = sr.getElementById('name');
         this.nameelement.innerHTML = this.name;
 
-        getUserStats(loyalty.getCookie('access_token'), (err, eventCount, pointsEarned) => {
-            // if user is not registered (user profile is not in database)
-            // create one for user
-            if (err == 'User is not registered') {
-                let mobileview = document.getElementById("mobileview");
-                mobileview.innerHTML = "";
-                let element = document.createElement('loading-spinner-element');
-                element.setAttribute("status", "User is marked for deletion...")
-                mobileview.appendChild(element)
+        if(this.mode=='INTEGRATED'){
 
-                setTimeout(() => {
-                    element.setAttribute("status", "Logging out...")
+            this.eventsattended = sr.getElementById('eventsattended');
+            this.eventsattended.innerHTML = this.events;
+            this.pointearned = sr.getElementById('pointearned');
+            this.pointearned.innerHTML = this.points;
+
+            getUserStats(loyalty.getCookie('access_token'), (err, eventCount, pointsEarned) => {
+                // if user is not registered (user profile is not in database)
+                // create one for user
+                if (err == 'User is not registered') {
+                    let mobileview = document.getElementById("mobileview");
+                    mobileview.innerHTML = "";
+                    let element = document.createElement('loading-spinner-element');
+                    element.setAttribute("status", "User is marked for deletion...")
+                    mobileview.appendChild(element)
+
                     setTimeout(() => {
-                        this.logout()
-                    }, 2500)
-                }, 2000)
-            }
-            if (eventCount != null) customElement.setAttribute('events', eventCount)
-            if (pointsEarned != null) customElement.setAttribute('points', pointsEarned)
-        })
+                        element.setAttribute("status", "Logging out...")
+                        setTimeout(() => {
+                            this.logout()
+                        }, 2500)
+                    }, 2000)
+                }
+                if (eventCount != null) customElement.setAttribute('events', eventCount)
+                if (pointsEarned != null) customElement.setAttribute('points', pointsEarned)
+            })
+        }
     }
 
     attributeChangedCallback(name, oldValue, newValue) {
-        console.log('account attribute changed callback')
+        console.log('ACCOUNT ATTRIBUTE CHANGED')
 
         if (name == 'events') {
             console.log('events changed')
